+++ app/pages/components/NextPage/NextPage.component.js
@@ -10,12 +10,13 @@ import { useNavigation } from '@react-navigation/native';
  * Pour créer une page :
  * - Ajoutez une page dans la fonction HomeStack de navigation
  * - Instanciez un composant NextPage comme cela :
- *  \<NextPage
- *        pageName = nomDeLaPageAAjouter
+ *  <NextPage
+ *        pageName = nomDeLaPageÀAjouter
  *        parameters = {varName:varValue}
  *        text = textAEcrireSurLeBouton
  *  />
 */
+
 class NextPage extends Component {
     constructor(props) {
         super(props);
@@ -63,16 +64,6 @@ class NextPage extends Component {
     }
 }
 
-/** 
- * Pour créer une page :
- * - Ajoutez une page dans la fonction HomeStack de navigation
- * - Instanciez un composant NextPage comme cela :
- *  \<NextPage
- *        pageName = nomDeLaPageAAjouter
- *        parameters = {varName:varValue}
- *        text = textAEcrireSurLeBouton
- *  />
-*/
 export default function (props) {
     const navigation = useNavigation();
     return <NextPage {...props} navigation={navigation} />;