+++ metro.config.js
@@ -12,7 +12,8 @@ defaultConfig.transformer = {
 };
 module.exports = defaultConfig;*/
 
-const { getDefaultConfig, mergeConfig } = require('@react-native/metro-config');
+const { getDefaultConfig } = require('expo/metro-config');
+const { mergeConfig } = require('@react-native/metro-config');
 
 const defaultConfig = getDefaultConfig(__dirname);