+++ CHANGELOG.md
@@ -4,5 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/), and this project adheres to [Semantic Versioning](https://semver.org/).
 
-<!-- ## Unreleased [{version_tag}](https://github.com/opengisch/qgis-plugin-ci/releases/tag/{version_tag}) - YYYY-MM-DD -->
+## Unreleased 
 
+### Added
+
+- topbarre text centered
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security